* commit '96137fb30f720f12a8ee93be589042900057d03d':
  Core/Misc: Future msvc warning fixes

+++ src/server/game/Spells/SpellInfo.cpp
@@ -413,7 +413,7 @@ bool SpellEffectInfo::IsEffect() const
 
 bool SpellEffectInfo::IsEffect(SpellEffectName effectName) const
 {
-    return Effect == uint32(effectName);
+    return Effect == effectName;
 }
 
 bool SpellEffectInfo::IsAura() const
